Update firebase login and add direct getOrder requests (#26)
* Implement ngx-auth-firebaseui, improve style

* Update firebase login and add direct getOrder requests

+++ src/app/my-account/my-account.component.ts
@@ -4,6 +4,8 @@ import { ConnectionService } from '../services/connection.service';
 import { FirebaseService } from '../services/firebase.service';
 import { WebsocketService } from '../services/websocket.service';
 
+import { MatDialog } from '@angular/material';
+import { DialogYesNoComponent } from '../dialogs/dialog-yes-no/dialog-yes-no.component';
 
 @Component({
   selector: 'app-my-account',
@@ -22,6 +24,7 @@ export class MyAccountComponent implements OnInit {
     public connectionService: ConnectionService,
     public firebaseService: FirebaseService,
     private wsService: WebsocketService,
+    public dialog: MatDialog,
   ) { }
 
   ngOnInit() {
@@ -29,7 +32,9 @@ export class MyAccountComponent implements OnInit {
 
   changeAlias(): void {
     if (this.wantsToChangeAlias) {
-      this.acceptAliasChange();
+      if (this.newAlias.trim() !== '') {
+        this.acceptAliasChange();
+      }
     } else {
       this.newAlias = this.connectionService.loggedInUser.alias;
       this.wantsToChangeAlias = true;
@@ -50,14 +55,29 @@ export class MyAccountComponent implements OnInit {
   }
 
   deleteMe(): void {
-    this.firebaseService.deleteUser()
-    .then(() => {
-      this.wsService.closeConnection();
-    })
-    .catch(error => {
-      if (error.code && error.code === 'auth/requires-recent-login') {
-        this.errorMessage = 'Your last login is too long back. ' +
-          'Please logoff + login again if you want to delete your account.';
+    const dialogRef = this.dialog.open(DialogYesNoComponent, {
+      width: '400px',
+      data: {
+        text: 'This will erase you permanently from the database and your peer list will be lost afterwards. Are you sure?',
+        yes: 'DELETE ME',
+        no: 'CANCEL',
+        yesColor: 'warn',
+        noColor: 'primary'
+      }
+    });
+
+    dialogRef.afterClosed().subscribe(isSure => {
+      if (isSure) {
+        this.firebaseService.deleteUser()
+        .then(() => {
+          this.wsService.closeConnection();
+        })
+        .catch(error => {
+          if (error.code && error.code === 'auth/requires-recent-login') {
+            this.errorMessage = 'Your last login is too long back. ' +
+              'Please logoff + login again if you want to delete your account.';
+          }
+        });
       }
     });
   }
